+++ .github/workflows/ci-master.yml
@@ -122,7 +122,11 @@ jobs:
         timeout-minutes: 40
         with:
           command: ci-test
-          args: --exclude=actix-redis --exclude=actix-session -- --nocapture
+          args: >-
+            --exclude=actix-redis
+            --exclude=actix-session
+            --exclude=actix-limitation
+            -- --nocapture
 
       - name: Clear the cargo caches
         run: |
