--- actix-limitation/src/core/mod.rs
@@ -1,3 +1,0 @@
-pub mod builder;
-pub mod errors;
-pub mod status;
